Fix memory corruption in stream decompression
There might be a situation after a call to qzDecompressStream() when
the stream's pending_out and pending_in are both greater than zero.
A user is supposed to repeatedly call the function until QATzip's output
buffer is empty. Then the user's input gets copied to the input buffer
until buffer_in is equal to the buffer's size.

The problem here is the data that is already in the buffer doesn't get
moved to the beginning of the buffer before appending new input to it.
As result the newly appended data gets copied beyond the input buffer's
boundary.

The proposed fix extends the stream's state with additional fields used
to flush the input buffer before copying new input to it.
Alternatively we can just shift the pending data to the begining of the
input buffer.